Refactor : Make engine.c adhere to its set of responsibilities
Function execute_atcmd in engine.c should execute the AT command once and parse the response to something user-readable.
Engine should be pointed to microcontroller- and/or compiler-specific function that handles reading of data received from UART module and not call the function directly.
These refactoring enhance modularity to the point that engine and at_parser can be used as libraries (as intended).

Changes:
- Removed for-loop in execute_atcmd which repeats sending of AT command if original response to AT command is a failure.
- Add function pointer to function handling receiving data from UART module as parameter to init_engine.
- Add static function that uses previously mentioned function pointer to copy contents of buffer used to store data received from UART module to a local buffer.

Resolves: #3